% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{bs_term}
\alias{bs_term}
\title{Term for a B-Spline Basis for Polynomial Splines}
\usage{
bs_term(
  x = numeric(),
  df = NULL,
  knots = NULL,
  degree = 3,
  intercept = FALSE,
  Boundary.knots = range(x)
)
}
\arguments{
\item{x, df, knots, degree, intercept, Boundary.knots}{same as \code{\link{bs}}.}
}
\description{
Term for a B-Spline Basis for Polynomial Splines
}
