% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_lb_gr}
\alias{joint_ms_lb_gr}
\title{Evaluates the Gradient of the Lower Bound}
\usage{
joint_ms_lb_gr(object, par, n_threads = object$max_threads)
}
\description{
Evaluates the Gradient of the Lower Bound
}
