% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_opt}
\alias{joint_ms_opt}
\title{Optimizes the Lower Bound}
\usage{
joint_ms_opt(
  object,
  par = object$start_val,
  rel_eps = 1e-08,
  max_it = 100L,
  n_threads = object$max_threads,
  c1 = 1e-04,
  c2 = 0.9,
  use_bfgs = TRUE,
  trace = 0L,
  cg_tol = 0.5,
  strong_wolfe = TRUE,
  max_cg = 0L,
  pre_method = 1L,
  quad_rule = object$quad_rule
)
}
\description{
Optimizes the Lower Bound
}
