% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_profile}
\alias{joint_ms_profile}
\title{Approximate Likelihood Ratio based Confidence Intervals}
\usage{
joint_ms_profile(
  object,
  opt_out,
  which_prof,
  delta,
  level = 0.95,
  max_step = 15L,
  rel_eps = 1e-08,
  max_it = 1000L,
  n_threads = object$max_threads,
  c1 = 1e-04,
  c2 = 0.9,
  use_bfgs = TRUE,
  trace = 0L,
  cg_tol = 0.5,
  strong_wolfe = TRUE,
  max_cg = 0L,
  pre_method = 3L,
  quad_rule = object$quad_rule,
  verbose = TRUE,
  mask = integer(),
  cache_expansions = object$cache_expansions,
  gr_tol = -1,
  hess = NULL
)
}
\arguments{
\item{object}{a joint_ms object from \code{\link{joint_ms_ptr}}.}

\item{opt_out}{maximum lower bound estimator from \code{\link{joint_ms_opt}}.}

\item{which_prof}{index of the parameter to profile.}

\item{delta}{numeric scalar greater than zero for the initial step size.
Steps are made of size \code{2^(iteration - 1) * delta}. A guess of the
standard deviation is a good value.}

\item{level}{confidence level.}

\item{max_step}{maximum number of steps to take in each direction when
constructing the approximate profile likelihood curve.}

\item{rel_eps, max_it, c1, c2, use_bfgs, trace, cg_tol, strong_wolfe, max_cg, pre_method, mask, gr_tol}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{n_threads}{number of threads to use.}

\item{quad_rule}{list with nodes and weights for a quadrature rule for the
integral from zero to one.}

\item{verbose}{logical for whether to print output during the construction
of the approximate profile likelihood curve.}

\item{cache_expansions}{\code{TRUE} if the expansions in the numerical
integration in the survival parts of the lower bound should be cached (not
recomputed). This requires more memory and may be an advantage
particularly with
expansions that take longer to compute (like \code{\link{ns_term}} and
\code{\link{bs_term}}). The computation time may be worse particularly if
you use more threads as the CPU cache is not well utilized.}

\item{hess}{the Hessian from \code{\link{joint_ms_hess}}. It is used to get
better starting values along the profile likelihood curve. Use \code{NULL}
if it is not passed.}
}
\description{
Approximate Likelihood Ratio based Confidence Intervals
}
