% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_ptr}
\alias{joint_ms_ptr}
\title{Creates a joint_ms Object to Estimate a Joint Survival and Marker Model}
\usage{
joint_ms_ptr(
  markers = list(),
  survival_terms = list(),
  max_threads = 1L,
  quad_rule = NULL,
  cache_expansions = TRUE
)
}
\arguments{
\item{markers}{either an object from \code{\link{marker_term}} or a list
of such objects.}

\item{survival_terms}{either an object from \code{\link{surv_term}} or a list
of such objects.}

\item{max_threads}{maximum number of threads to use.}

\item{quad_rule}{list with nodes and weights for a quadrature rule for the
integral from zero to one.}

\item{cache_expansions}{\code{TRUE} if the expansions in the numerical
integration in the survival parts of the lower bound should be cached (not
recomputed). This requires more memory and may be an advantage
particularly with
expansions that take longer to compute (like \code{\link{ns_term}} and
\code{\link{bs_term}}). The computation time may be worse particularly if
you use more threads as the CPU cache is not well utilized.}
}
\description{
Creates a joint_ms Object to Estimate a Joint Survival and Marker Model
}
