% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_set_vcov}
\alias{joint_ms_set_vcov}
\title{Sets the Covariance Parameters}
\usage{
joint_ms_set_vcov(
  object,
  vcov_vary,
  vcov_surv,
  par = object$start_val,
  va_mean = NULL
)
}
\arguments{
\item{object}{a joint_ms object from \code{\link{joint_ms_ptr}}.}

\item{vcov_vary}{the covariance matrix for the time-varying effects.}

\item{vcov_surv}{the covariance matrix for the frailties.}

\item{par}{parameter vector to be formatted.}

\item{va_mean}{a matrix with the number of rows equal to the number of
random effects per observation and the number of columns is the number
of observations. The order for the observations needs to be the same as the
\code{id} element of \code{object}.}
}
\description{
Sets the covariance matrices to the passed values. The function also sets
covariance matrices for the variational distributions to the same values.
}
