% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{joint_ms_va_par}
\alias{joint_ms_va_par}
\title{Extracts the Variational Parameters}
\usage{
joint_ms_va_par(object, par = object$start_val)
}
\arguments{
\item{object}{a joint_ms object from \code{\link{joint_ms_ptr}}.}

\item{par}{parameter vector to be formatted.}
}
\value{
A list with one list for each individual with the estimated mean and
covariance matrix.
}
\description{
Computes the estimated variational parameters for each individual.
}
