% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker-term.R
\name{marker_term}
\alias{marker_term}
\title{Creates Data for One Type of Marker}
\usage{
marker_term(formula, id, data, time_fixef, time_rng)
}
\arguments{
\item{formula}{a two-sided \code{\link{formula}} with the marker outcome
on the left-hand side and fixed effect covariates on the right-hand
side.}

\item{id}{the variable for the id of each individual.}

\item{data}{a \code{data.frame} or environment to look at up the
variables in.}

\item{time_fixef}{the time-varying fixed effects. See .e.g.
\code{\link{poly_term}}.}

\item{time_rng}{the time-varying random effects. See .e.g.
\code{\link{poly_term}}.}
}
\description{
Creates Data for One Type of Marker
}
\details{
The \code{time_fixef} should likely not include an intercept as this is
often included in \code{formula}. Use
\code{poly_term(degree = 0, raw = TRUE, intercept = TRUE)} if you want only
a random intercept.
}
