% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{ns_term}
\alias{ns_term}
\title{Term for a Basis Matrix for Natural Cubic Splines}
\usage{
ns_term(
  x = numeric(),
  df = NULL,
  knots = NULL,
  intercept = FALSE,
  Boundary.knots = range(if (use_log) log(x) else x),
  use_log = FALSE
)
}
\arguments{
\item{x, df, knots, intercept, Boundary.knots}{same as \code{\link{ns}}.}

\item{use_log}{\code{TRUE} if the polynomials should be in the log of the
argument.}
}
\description{
Term for a Basis Matrix for Natural Cubic Splines
}
