% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{stacked_term}
\alias{stacked_term}
\title{Term for a Basis Matrix for of Different Types of Terms}
\usage{
stacked_term(...)
}
\arguments{
\item{...}{term objects from the package.}
}
\description{
Creates a basis matrix consisting of different types of terms.
E.g. to create a varying-coefficient.
}
\examples{
# TODO: add an example

}
\seealso{
\code{\link{poly_term}}, \code{\link{bs_term}}, \code{\link{ns_term}}, and
\code{\link{weighted_term}}
}
