% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_term.R
\name{surv_term}
\alias{surv_term}
\title{Creates Data for One Type of Survival Outcome}
\usage{
surv_term(formula, id, data, time_fixef, ders = NULL, with_frailty = FALSE)
}
\arguments{
\item{formula}{a two-sided \code{\link{formula}} with the survival outcome
on the left-hand side and fixed effect covariates on the right-hand
side. The left-hand side needs to be a \code{\link{Surv}} object and can
be either right-censored and left-truncated.}

\item{id}{the variable for the id of each individual.}

\item{data}{\code{\link{data.frame}} with at least the time variable.}

\item{time_fixef}{the time-varying fixed effects. See .e.g.
\code{\link{poly_term}}. This is for the baseline hazard. Note that many
basis expansions has boundary knots. It is important that these are set
to cover the full range of survival times including time zero for some
expansions.}

\item{ders}{a \code{\link{list}} with \code{\link{integer}} vectors for how
the survival outcome is linked to the markers. 0 implies present values,
-1 is integral of, and 1 is the derivative. \code{NULL} implies the present
value of the random effect for all markers.}

\item{with_frailty}{\code{TRUE} if there should be a frailty term.}
}
\description{
Creates Data for One Type of Survival Outcome
}
\details{
The \code{time_fixef} should likely not include an intercept as this is
often included in \code{formula}.
}
