% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{VAJointSurv-terms}
\alias{VAJointSurv-terms}
\title{Expansions in the VAJointSurv package}
\description{
The VAJointSurv package uses different functions to allow for expansions in
time possibly with covariate interactions. The main usage of the functions
is internally but they do provide an element called `eval()` which is a
function to evaluate the expansion. These functions take the following
arguments:

\itemize{
\item \code{x} numeric vector with points at which to evaluate the expansion.
\item \code{der} integer indicating whether to evaluate the expansion, its integral,
or the derivative.
\item \code{lower_limit} possible lower limit if integration is performed.
\item \code{newdata} a \code{\link{data.frame}} with new data if this is required.
E.g. for \code{\link{weighted_term}}.
}

The supported terms are \code{\link{ns_term}}, \code{\link{bs_term}},
\code{\link{poly_term}}, \code{\link{weighted_term}}, and a
\code{\link{stacked_term}}.
}
