% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{weighted_term}
\alias{weighted_term}
\title{Term for a Basis Matrix for Weighted Term}
\usage{
weighted_term(x, weight)
}
\arguments{
\item{x}{a term type from the package.}

\item{weight}{a numeric vector with weights for x.}
}
\description{
Creates a weighted basis matrix where the entries are weighted with a
numeric vector to e.g. create a varying-coefficient.
}
\examples{
# TODO: add an example

}
\seealso{
\code{\link{poly_term}}, \code{\link{bs_term}}, \code{\link{ns_term}}, and
\code{\link{stacked_term}}
}
