% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_profile}
\alias{joint_ms_profile}
\title{Approximate Likelihood Ratio based Confidence Intervals}
\usage{
joint_ms_profile(
  object,
  opt_out,
  which_prof,
  delta,
  level = 0.95,
  max_step = 15L,
  rel_eps = 1e-08,
  max_it = 100L,
  n_threads = object$max_threads,
  c1 = 1e-04,
  c2 = 0.9,
  use_bfgs = TRUE,
  trace = 0L,
  cg_tol = 0.5,
  strong_wolfe = TRUE,
  max_cg = 0L,
  pre_method = 1L,
  quad_rule = object$quad_rule,
  verbose = TRUE
)
}
\arguments{
\item{object}{a joint_ms object from \code{\link{joint_ms_ptr}}.}

\item{opt_out}{maximum lower bound estimator from \code{\link{joint_ms_opt}}.}

\item{which_prof}{index of the parameter to profile.}

\item{delta}{numeric scalar greater than zero for the initial step size.
Steps are made of size \code{2^(iteration - 1) * delta}. A guess of the
standard deviation is a good value.}

\item{level}{confidence level.}

\item{max_step}{maximum number of steps to take in each direction when
constructing the approximate profile likelihood curve.}

\item{rel_eps}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{max_it}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{n_threads}{number of threads to use.}

\item{c1}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{c2}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{use_bfgs}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{trace}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{cg_tol}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{strong_wolfe}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{max_cg}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{pre_method}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{quad_rule}{list with nodes and weights for a quadrature rule for the
integral from zero to one.}

\item{verbose}{logical for whether to print output during the construction
of the approximate profile likelihood curve.}
}
\description{
Approximate Likelihood Ratio based Confidence Intervals
}
