% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_ptr}
\alias{joint_ms_ptr}
\title{Creates a joint_ms to Estimate a Joint Survival and Marker Model}
\usage{
joint_ms_ptr(markers, max_threads = 1L)
}
\description{
Creates a joint_ms to Estimate a Joint Survival and Marker Model
}
