% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_set_vcov}
\alias{joint_ms_set_vcov}
\title{Sets the Covariance Parameters}
\usage{
joint_ms_set_vcov(object, vcov_vary, vcov_surv, par = object$start_val)
}
\arguments{
\item{object}{a joint_ms object from \code{\link{joint_ms_ptr}}.}

\item{vcov_vary}{the covariance matrix for the time-varying effects.}

\item{vcov_surv}{the covariance matrix for the frailties.}

\item{par}{parameter vector to be formatted.}
}
\description{
Sets the covariance matrices to the passed value. The function also sets
covariance matrices for the variational distributions to the same value.
}
