% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_start_val}
\alias{joint_ms_start_val}
\title{Quick Heuristic for the Starting Values}
\usage{
joint_ms_start_val(
  object,
  n_threads = object$max_threads,
  rel_eps = 1e-08,
  c1 = 1e-04,
  c2 = 0.9,
  max_it = 100L,
  quad_rule = object$quad_rule
)
}
\description{
Quick Heuristic for the Starting Values
}
