% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_start_val}
\alias{joint_ms_start_val}
\title{Quick Heuristic for the Starting Values}
\usage{
joint_ms_start_val(
  object,
  par = object$start_val,
  rel_eps = 1e-08,
  max_it = 1000L,
  n_threads = object$max_threads,
  c1 = 1e-04,
  c2 = 0.9,
  use_bfgs = TRUE,
  trace = 0,
  cg_tol = 0.5,
  strong_wolfe = TRUE,
  max_cg = 0,
  pre_method = 3L,
  quad_rule = object$quad_rule,
  mask = integer(),
  cache_expansions = object$cache_expansions,
  gr_tol = -1
)
}
\arguments{
\item{object}{a joint_ms object from \code{\link{joint_ms_ptr}}.}

\item{par}{starting value.}

\item{rel_eps, c1, c2, max_it, gr_tol}{arguments to pass to the C++ version of
\code{\link{psqn}}.}

\item{n_threads}{number of threads to use.}

\item{use_bfgs}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{trace}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{cg_tol}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{strong_wolfe}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{max_cg}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{pre_method}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{quad_rule}{list with nodes and weights for a quadrature rule for the
integral from zero to one.}

\item{mask}{arguments to pass to the C++ version of \code{\link{psqn}}.}

\item{cache_expansions}{\code{TRUE} if the expansions in the numerical
integration in the survival parts of the lower bound should be cached (not
recomputed). This requires more memory and may be an advantage
particularly with
expansions that take longer to compute (like \code{\link{ns_term}} and
\code{\link{bs_term}}). The computation time may be worse particularly if
you use more threads as the CPU cache is not well utilized.}
}
\description{
Quick Heuristic for the Starting Values
}
