% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_surv_VA.R
\name{joint_ms_start_val}
\alias{joint_ms_start_val}
\title{Quick Heuristic for the Starting Values}
\usage{
joint_ms_start_val(
  object,
  n_threads = object$max_threads,
  rel_eps = 1e-08,
  c1 = 1e-04,
  c2 = 0.9,
  max_it = 100L,
  quad_rule = object$quad_rule
)
}
\arguments{
\item{object}{a joint_ms object from \code{\link{joint_ms_ptr}}.}

\item{n_threads}{number of threads to use.}

\item{rel_eps, c1, c2, max_it}{arguments to pass to the C++ version of
\code{\link{psqn}}.}

\item{quad_rule}{list with nodes and weights for a quadrature rule for the
integral from zero to one.}
}
\description{
Quick Heuristic for the Starting Values
}
