% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{ns_term}
\alias{ns_term}
\title{Term for a Basis Matrix for Natural Cubic Splines}
\usage{
ns_term(
  x = numeric(),
  df = NULL,
  knots = NULL,
  intercept = FALSE,
  Boundary.knots = range(x)
)
}
\arguments{
\item{x, df, knots, intercept, Boundary.knots}{same as \code{\link{ns}}.}
}
\description{
Term for a Basis Matrix for Natural Cubic Splines
}
