% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_marker}
\alias{plot_marker}
\title{Plots a Markers Mean Curve with Pointwise Quantiles}
\usage{
plot_marker(
  time_fixef,
  time_rng,
  fixef_vary,
  x_range,
  vcov_vary,
  p = 0.95,
  xlab = "Time",
  ylab = "Marker",
  newdata = NULL,
  ...
)
}
\arguments{
\item{time_fixef}{the time-varying fixed effects. See .e.g.
\code{\link{poly_term}}.}

\item{time_rng}{the time-varying random effects. See .e.g.
\code{\link{poly_term}}.}

\item{fixef_vary}{fixed effect coefficients for \code{time_fixef}.}

\item{x_range}{2D numeric vector with start and end points.}

\item{vcov_vary}{the covariance matrix for \code{time_rng}.}

\item{p}{coverage of the two quantiles.}

\item{xlab, ylab, ...}{arguments passed to \code{\link{plot}}.}

\item{newdata}{\code{data.frame} with data for the weights if any.}
}
\description{
Plots a Markers Mean Curve with Pointwise Quantiles
}
