% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_surv}
\alias{plot_surv}
\title{Plots Quantilies of the Hazard}
\usage{
plot_surv(
  time_fixef,
  time_rng,
  x_range,
  vcov_vary,
  Xi_var,
  ps = c(0.025, 0.5, 0.975),
  log_hazard_shift = 0,
  xlab = "Time",
  ylab = "Marker",
  ...
)
}
\description{
Plots Quantilies of the Hazard
}
