% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{poly_term}
\alias{poly_term}
\title{Term for Orthogonal Polynomials}
\usage{
poly_term(
  x = numeric(),
  degree = 1,
  coefs = NULL,
  raw = FALSE,
  intercept = FALSE
)
}
\arguments{
\item{x, degree, coefs, raw}{same as \code{\link{poly}}.}

\item{intercept}{\code{TRUE} if there should be an intercept.}
}
\description{
Term for Orthogonal Polynomials
}
