% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_term.R
\name{surv_term}
\alias{surv_term}
\title{Creates Data for One Type of Survival Outcome}
\usage{
surv_term(formula, id, data, time_fixef)
}
\arguments{
\item{formula}{a two-sided \code{\link{formula}} with the survival outcome
on the left-hand side and fixed effect covariates on the right-hand
side. The left-hand side needs to be a \code{\link{Surv}} object and can
be either right-censored and left-truncated.}

\item{id}{the variable for the id of each individual.}

\item{data}{a \code{data.frame} or environment to look at up the
variables in.}

\item{time_fixef}{the time-varying fixed effects. See .e.g.
\code{\link{poly_term}}. This is for baseline hazard.}
}
\description{
Creates Data for One Type of Survival Outcome
}
\details{
The \code{time_fixef} should likely not include an intercept as this is
often included in \code{formula}.
}
